Tidy up code and small code changes
