+++ proximity-feature/src/main/java/eu/europa/ec/proximityfeature/interactor/ProximitySuccessInteractor.kt
@@ -21,7 +21,7 @@ import eu.europa.ec.businesslogic.extension.safeAsync
 import eu.europa.ec.commonfeature.ui.document_details.transformer.DocumentDetailsTransformer.toListItemData
 import eu.europa.ec.commonfeature.ui.document_details.transformer.transformToDocumentDetailsDocumentItem
 import eu.europa.ec.commonfeature.ui.document_success.model.DocumentSuccessItemUi
-import eu.europa.ec.commonfeature.ui.request.transformer.toPath
+import eu.europa.ec.commonfeature.ui.request.transformer.toClaimPath
 import eu.europa.ec.corelogic.controller.WalletCoreDocumentsController
 import eu.europa.ec.corelogic.controller.WalletCorePresentationController
 import eu.europa.ec.corelogic.extension.getLocalizedClaimName
@@ -78,7 +78,7 @@ class ProximitySuccessInteractorImpl(
                     val detailsDocumentItems = document.data.claims
                         .filter { claim ->
                             disclosedDocument.disclosedItems.any { disclosedItem ->
-                                claim.identifier == disclosedItem.toPath().last() //TODO check this
+                                claim.identifier == disclosedItem.toClaimPath().value.last() //TODO check this
                             }
                         }
                         .map { claim ->
