+++ ui-logic/src/main/java/eu/europa/ec/uilogic/component/wrap/WrapListItems.kt
@@ -68,7 +68,7 @@ fun WrapListItems(
 
                 ListItem(
                     modifier = itemModifier,
-                    item = item.collapsed,
+                    item = item.header,
                     onItemClick = onItemClick,
                     hideSensitiveContent = hideSensitiveContent,
                     mainContentVerticalPadding = mainContentVerticalPadding,
